+++ manutencao-spring/manutencao/src/main/java/br/net/dwii/projeto/manutencao/model/dao/DaoI.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface DaoI<T> {
     public void add(T objeto) throws Exception; 
     public List<T> getAll() throws Exception;
-    public T getById(long id) throws Exception;
+    public T getById(int id) throws Exception;
     public void update(T objeto) throws Exception;
     public void delete(T objeto) throws Exception;
 }
